// src/lib/models/DisciplineType.ts
export enum DisciplineType {
  PHYSIOTHERAPY = 'PHYSIOTHERAPY',
  OCCUPATIONAL_THERAPY = 'OCCUPATIONAL_THERAPY',
  SPEECH_THERAPY = 'SPEECH_THERAPY',
  CHIROPRACTIC = 'CHIROPRACTIC',
  PODIATRY = 'PODIATRY',
  DIETETICS = 'DIETETICS',
  PSYCHOLOGY = 'PSYCHOLOGY',
  OSTEOPATHY = 'OSTEOPATHY',
  AUDIOLOGY = 'AUDIOLOGY',
  GENERAL = 'GENERAL'
}